refactor: move folders to src level
