+++ web-frontend/src/public/favicon.ico
