+++ web-frontend/src/styles/globals.scss
